requirements.txt
pandas
numpy
faiss-cpu
torch
tqdm
sentence-transformers
scikit-learn
boto3
